Se aplico el modelo que se diseno para generar los informes, a traves de mejoras a las diferentes clases enfocadas en la clase usuario y egreso

+++ src/main/java/com/example/repository/UsuarioCRUDRepository.java
@@ -3,6 +3,6 @@ package com.example.repository;
 import com.example.entities.Usuario;
 import org.springframework.data.repository.CrudRepository;
 
-public interface UsuarioCRUDRepository extends CrudRepository<Usuario,Integer> {
+public interface UsuarioCRUDRepository extends CrudRepository<Usuario,Long> {
     
 }
